requirements.txt
python-dotenv
pymysql
pandas
boto3
pytest 
moto 
pytest-mock
tqdm